backend-offline/requirements.txt
fastapi==0.116.2
uvicorn==0.35.0
torch==2.6.0+cu118
torchvision==0.21.0+cu118
pillow==10.2.0
numpy==1.26.4
python-multipart==0.0.20